+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+* Chore: added tests for nodes.
+* Fix: recursive `toJson()` calls in `NodeDetails`.
+* Fix: bad type comparations in `owner` setter into `Node` class.
+
 ## 1.0.1
 
 * Fix: Node levels not updating on insertion/addition [#1](https://github.com/Novident/novident-nodes/pull/1).